[GLUTEN-7031][CORE] Move JNI / Exception utilities to gluten-core
