+++ backends-velox/src/main/scala/org/apache/gluten/utils/SharedLibraryLoaderCentos9.scala
@@ -16,7 +16,7 @@
  */
 package org.apache.gluten.utils
 
-import org.apache.gluten.vectorized.JniLibLoader
+import org.apache.gluten.jni.JniLibLoader
 
 class SharedLibraryLoaderCentos9 extends SharedLibraryLoader {
   override def loadLib(loader: JniLibLoader): Unit = {
