+++ gluten-core/src/main/java/org/apache/gluten/jni/JniLibLoader.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.vectorized;
+package org.apache.gluten.jni;
 
 import org.apache.gluten.exception.GlutenException;
