+++ gluten-core/src/main/java/org/apache/spark/util/GlutenShutdownManager.scala
